requirements.txt
django==3.0.8
django-tinymce==3.0.2
django-crispy-forms==1.9.1
xhtml2pdf==0.2.4
python-gettext==4.0